removing lh_at_entry after calculating first and last lh date. Once we use first and last lh date to trim the EntryDate, by construction, all enrollments are LH at entry. So there's no need for that variable.
Modifying how we determine if an LH date is in the start window - let's use the same in-window calculations we use elsewhere, rather than trying to needlessly recreate the wheel.

Also modifying the was_lh_at_end and was_lh_during_period vars to do the same look across ALL LH dates for the person, not just the EECR's/LECR's.